add global test

+++ oneflow/core/functional/impl/array_functor.cpp
@@ -2129,7 +2129,7 @@ class SplitFunctor {
 class UnbindFunctor {
  public:
   UnbindFunctor() {}
-  Maybe<TensorTuple> operator()(const std::shared_ptr<one::Tensor>& x, const int32_t& dim) const {
+  Maybe<TensorTuple> operator()(const std::shared_ptr<one::Tensor>& x, const int64_t& dim) const {
     int32_t axis = dim;
     const int32_t ndim = x->ndim();
     if (axis < 0) { axis += ndim; }
